fixup! Address PR comments. Move tests into GcsUtilTest instead of using the pipeline.
